# perf: date formatting optimization