--- a/src/libs/ReportUtils.ts
+++ b/src/libs/ReportUtils.ts
@@ -222,7 +222,6 @@ import {
     getCurrency,
     getDescription,
     getFormattedCreated,
-    getFormattedPostedDate,
     getMCCGroup,
     getMerchant,
     getMerchantOrDescription,