--- a/src/libs/TransactionUtils/index.ts
+++ b/src/libs/TransactionUtils/index.ts
@@ -1735,6 +1735,7 @@ export {
     getMCCGroup,
     getCreated,
     getFormattedCreated,
+    getPostedDate,
     getCategory,
     getBillable,
     getTag,