--- a/src/libs/actions/IOU.ts
+++ b/src/libs/actions/IOU.ts
@@ -11726,6 +11726,7 @@ function initSplitExpense(transaction: OnyxEntry<OnyxTypes.Transaction>, isOpenC
 
     const transactionDetails = getTransactionDetails(transaction);
     const transactionDetailsAmount = transactionDetails?.amount ?? 0;
+    const formattedCreated = transactionDetails?.created ? DateUtils.formatWithUTCTimeZone(transactionDetails.created, CONST.DATE.FNS_FORMAT_STRING) : undefined;
     const defaultCreated = DateUtils.formatWithUTCTimeZone(DateUtils.getDBTime(), CONST.DATE.FNS_FORMAT_STRING);
 
     const draftTransaction = buildOptimisticTransaction({
@@ -11738,7 +11739,7 @@ function initSplitExpense(transaction: OnyxEntry<OnyxTypes.Transaction>, isOpenC
                     description: transactionDetails?.comment,
                     category: transactionDetails?.category,
                     tags: transaction?.tag ? [transaction?.tag] : [],
-                    created: transactionDetails?.created ?? defaultCreated,
+                    created: formattedCreated ?? defaultCreated,
                 },
                 {
                     transactionID: NumberUtils.rand64(),
@@ -11746,7 +11747,7 @@ function initSplitExpense(transaction: OnyxEntry<OnyxTypes.Transaction>, isOpenC
                     description: transactionDetails?.comment,
                     category: transactionDetails?.category,
                     tags: transaction?.tag ? [transaction?.tag] : [],
-                    created: transactionDetails?.created ?? defaultCreated,
+                    created: formattedCreated ?? defaultCreated,
                 },
             ],
             amount: transactionDetailsAmount,
@@ -11807,6 +11808,7 @@ function addSplitExpenseField(transaction: OnyxEntry<OnyxTypes.Transaction>, dra
     }
 
     const transactionDetails = getTransactionDetails(transaction);
+    const formattedCreated = transactionDetails?.created ? DateUtils.formatWithUTCTimeZone(transactionDetails.created, CONST.DATE.FNS_FORMAT_STRING) : undefined;
 
     Onyx.merge(`${ONYXKEYS.COLLECTION.SPLIT_TRANSACTION_DRAFT}${transaction.transactionID}`, {
         comment: {
@@ -11818,7 +11820,7 @@ function addSplitExpenseField(transaction: OnyxEntry<OnyxTypes.Transaction>, dra
                     description: transactionDetails?.comment,
                     category: transactionDetails?.category,
                     tags: transaction?.tag ? [transaction?.tag] : [],
-                    created: transactionDetails?.created ?? DateUtils.formatWithUTCTimeZone(DateUtils.getDBTime(), CONST.DATE.FNS_FORMAT_STRING),
+                    created: formattedCreated ?? DateUtils.formatWithUTCTimeZone(DateUtils.getDBTime(), CONST.DATE.FNS_FORMAT_STRING),
                 },
             ],
         },
@@ -11853,13 +11855,14 @@ function updateSplitExpenseField(splitExpenseDraftTransaction: OnyxEntry<OnyxTyp
     const splitExpenses = draftTransaction?.comment?.splitExpenses?.map((item) => {
         if (item.transactionID === splitExpenseTransactionID) {
             const transactionDetails = getTransactionDetails(splitExpenseDraftTransaction);
+            const formattedCreated = transactionDetails?.created ? DateUtils.formatWithUTCTimeZone(transactionDetails.created, CONST.DATE.FNS_FORMAT_STRING) : undefined;
 
             return {
                 ...item,
                 description: transactionDetails?.comment,
                 category: transactionDetails?.category,
                 tags: splitExpenseDraftTransaction?.tag ? [splitExpenseDraftTransaction?.tag] : [],
-                created: transactionDetails?.created ?? DateUtils.formatWithUTCTimeZone(DateUtils.getDBTime(), CONST.DATE.FNS_FORMAT_STRING),
+                created: formattedCreated ?? DateUtils.formatWithUTCTimeZone(DateUtils.getDBTime(), CONST.DATE.FNS_FORMAT_STRING),
             };
         }
         return item;